+++ wiki/Changelog.md
@@ -1,6 +1,9 @@
 Font rendering
   * Exclude full-cell characters from auto-narrowing (#918).
 
+Configuration
+  * Emojis - Placement values translatable (#919).
+
 ### 3.0.3 (28 September 2019) ###
 
 Terminal features